159: Move types to one file, change outout from transformJson
