+++ frontend/src/routes/report/page.tsx
@@ -13,35 +13,11 @@ import {
 } from '@mantine/core';
 import { useForm } from '@mantine/form';
 import { useEffect, useState } from 'react';
-import { RowWrapper } from './inputRow.tsx';
-import {
-  AnalysisTypeSchema,
-  DataSource,
-  DataSourceSchema,
-  FeedKindSchema,
-  FeedSortingKindSchema,
-  FeedSortingTimeSchema,
-} from '../../rmoods/client/types.ts';
+import { DataSource } from '../../rmoods/client/types.ts';
 import { zodResolver } from 'mantine-form-zod-resolver';
-import { z } from 'zod';
 import { DataSourceTable } from './tables.tsx';
 import { transformJson } from './transformJson.ts';
-
-const formValidationSchema = z.object({
-  name: z.string().min(1, { message: 'Name must be longer than 1 character' }),
-  resourceType: FeedKindSchema,
-  isPublic: z.enum(['true', 'false']),
-  size: z.enum(['small', 'medium', 'large', 'custom']),
-  customSize: z.number().optional(),
-  sortBy: FeedSortingKindSchema,
-  time: FeedSortingTimeSchema,
-  dataSource: z
-    .array(DataSourceSchema)
-    .min(1, { message: 'At least 1 data source is required' }),
-  analyses: AnalysisTypeSchema,
-});
-
-export type formValues = z.infer<typeof formValidationSchema>;
+import { formValidationSchema, formValues, RowWrapper } from './types.ts';
 
 /**
  * Report component for creating a new report.
